Update Solve The Maze.cpp

+++ 1700/Solve The Maze.cpp
@@ -19,8 +19,8 @@ int main(){
         {
 		cin>>n>>m;
 		for(ll i=1;i<=n;i++)scanf("%s",a[i]+1);
-		for(int i=0;i<=n+1;i++) for(int j=0;j<=m+1;j++) b[i][j]=v[i][j]=0;
-		for(int i=1;i<=n;i++) for(int j=1;j<=m;j++){
+		for(ll i=0;i<=n+1;i++) for(int j=0;j<=m+1;j++) b[i][j]=v[i][j]=0;
+		for(ll i=1;i<=n;i++) for(int j=1;j<=m;j++){
 		   if(a[i][j]=='#') b[i][j]=1;
 		   if(a[i][j]=='B') b[i][j]=b[i-1][j]=b[i+1][j]=b[i][j-1]=b[i][j+1]=1;
 	   }
